addon: moved synchronization from server to addon package, migrated code from AddonManager into it
